% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_calc.R
\name{my_calc}
\alias{my_calc}
\title{A calculator function}
\usage{
my_calc(x, y, fun)
}
\arguments{
\item{x}{A vector or matrix}

\item{y}{A vector or matrix}

\item{fun}{A string}
}
\description{
This function allows you to multiply, divide, add, or
subtract two inputted numbers.
}
\examples{
x <- 1:5
y <- c(5, 2, 3, 10, 2)
fun <- "divide"
my_calc(x, y, "plus")
my_calc(x, y, "subtract")
my_calc(x, y, fun)
my_calc(x, y, "multiply")
}
