% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_multiplier.R
\name{my_multiplier}
\alias{my_multiplier}
\title{A number-multiplying function}
\usage{
my_multiplier(value)
}
\arguments{
\item{value}{A number}
}
\description{
This function multiplies numbers by specified amounts based on
specified criteria. Values below 5 are unchanged, values equal to or above 5
and below 10 are multiplied by 5, and values 10 or higher are mulitplied by
10.
}
\examples{
my_multiplier(1)
my_multiplier(7)
my_multiplier(11)
}
